Enable setting the Z layer for 2D particles
Add a new field `EffectAsset::z_layer_2d` holding the Z coordinate of
the layer at which particles are rendered in 2D mode. Also add a new
field `ParticleEffect::z_layer_2d` to override this value on a
per-instance basis.

Reflect both `EffectAsset` and `ParticleEffect`, allowing to tweak the
`z_layer_2d` value on a per-instance basis with the egui inspector in
the `2d.rs` example, to show the effect of this Z layer feature.

Bug: #28