Single bind group for particles and indirection
Use a single common bind group for both the particle buffer containing
the actual per-particle data, and the indirect buffer containing the
index of the particles to render in sorted order. This frees up a bind
group for other uses in the update compute shader, while keeping their
number under the wgpu default limit of 4.