Fix same-frame despawn/respawn (#107)
Fix bind groups not re-created correctly when an effect is despawned in
the same frame as another effect is spawned.

Add a partial workaround for the `vfx_indirect.wgsl` shader accessing
GPU buffers in order from index 0, which requires all GPU particle
buffers to be in use without any gap. The partial workaround first
destroys buffers for de-allocated effects, before re-creating new
buffers for newly-spawned ones. This is not perfect, because this
doesn't account for gaps when more buffers are de-allocated than created
in the same frame, in which case `vfx_indirect` accesses the first N
buffers, but some of them are not in use and some in use are skipped.
We're missing an indirection here.

Fixes #106