Dynamic attribute-based particle layout (#129)
This change introduces a new dynamic particle layout based on _attributes_. A
particle attribute is a single scalar or vector value describing one aspect of
the particle. The value is stored per particle, and the set of attributes for
all the particles of an effect is organized into a layout describing the offset
and alignment of each attribute into the particle buffer on the GPU.

This new design allows customizing the particle storage itself based on the
modifiers used for each effect, which enables both greater customizing and a
compact memory footprint, as particles only store the attributes needed by the
effect.

To enable even greater customization, _properties_ can be attached to effects.
A property is a dynamic quantity assignable at runtime, and uploaded to the GPU
each frame, allowing to control the quantity it's bound to. Properties can be
bound to values of modifiers, like the `AccelModifer` acceleration, to enable
users to vary that field at runtime without the need to recreate the effect.
Properties trade some performance for greater control and customizing. This
generalize an implicit concept previously only available for the acceleration
of `AccelModifier` and the force field sources, and now made available to any
property-enabled field of any modifier.

Fixes #109