Add `AabbKillModifier`
Add a new update modifier `AabbKillModifier` which kills all particles
entering or exiting its AABB. This modifier is useful to confine the
particles to a space (_e.g._ a room) and avoid leaking into adjacent
areas, or ensuring no particle enter a defined space.

Update the force field example to use that new `AabbKillModifier` to
define an "allow" box (green) and a "forbid" box (red) demonstrating how
the modifier works.

Incidentally, this change also fixes mouse clicks in the
`force_field.rs` example when the window is resized.

Partially fixes #118