Split `PositionCircleModifer` and add tangent
Split the `PositionCircleModifer` into a position part and a velocity
part. The former becomes the `InitPositionCircleModifier` while the
latter is the new `InitVelocityCircleModifier`.

Add a new `InitVelocityTangentModifier` initializing the velocity of particles
along the tangent to a cylinder. Add a new example `portal.rs` demonstrating
the use of that new modifier.

Related to #97