Fixed PositionCone3dModifier placing particles at twice the translation
The position inside the cone was affected by the transform, but since the w component of the position was set to 1 the transform was affecting adding the translation.
However this also gets done when it's rendered I believe, so the translation was added twice.

However I'm not sure why the position (and the speed in the next part) were affected by the transform at all. The code for InitPositionCircleModifier does not use transform, though the InitVelocityCircleModifier does, so maybe this change is the most correct?
I can't seem to see any observable difference at all whether I keep it or remove it entirely, but maybe I'm missing something.
Though I think maybe neither of them should be affected by the transform? But maybe I'm missing something here.

Fixes #152