Add `MotionIntegration`
Add a new `MotionIntegration` setting to `EffectAsset` to configure
when, if at all, the motion integration of particles occurs relative to
the application of modifiers. During the update pass, the optional
motion integration code will be injected before or after the code for
all modifiers. The default value is `MotionIntegration::PostUpdate`
corresponding to the previous behavior of integrating the motion after
the modifiers were applied.