Split runtime data into separate component (#176)
Split the runtime data generated by the library from the user-provided
data, and store the former into a new `CompiledParticleEffect`
component. This enables leveraging the Bevy change detection mechanism
to invalidate the cached data associated with the effect instance.

Also split the spawner configuration into a separate `EffectSpawner`
component. The spawner values of the `EffectAsset` and `ParticleEffect`
are used to initialize this new component, which holds the runtime data
for the spawner.

Fixes #162