Rename `spawn::Value` to `CpuValue` and fix `Hash`
Rename the `spawn::Value` type to `CpuValue` to prevent confusion with
the more recent `graph::Value`, which is destined to eventually replace
it.

Add `impl Hash for CpuValue` for all floating-point types via a helper
trait `FloatHash`. This enables deriving `Hash` for types using
`CpuValue<T: FloatHash>`, which at the time of writing includes all
uses.

Fix the `Hash` implementation of `SetColorModifier` and
`SetSizeModifier`, which were ignoring the `CpuValue` enum variant.