Multi-context modifiers with expression API
Finish migrating all modifiers and all examples to use the Expression
API.

Change `ModifierContext` to a bitflag and allow modifiers to be used in
multiple contexts. Transition some of them like `SetAttributeModifier`
to be useable in both init and update contexts. Rename the apply
functions, qualifying them by context (_e.g._ `apply_update()` for the
apply function of the `UpdateModifier` trait).

Reorganize the `modifier` module to group modifiers by category, instead
of context. This incidentally resolve a nightly-only bug where the
`modifier/render` module was being re-exported publicly, shadowing the
private `render` module and causing a nightly build failure.

Add a new `KillSphereModifier` and rename the other one to
`KillAabbModifier` for consistency.

Remove a few unused types and modifiers, and the `resolve_properties()`
mechanism of the `Modifier` trait, made obsolete by the Expression API.