Add local space simulation and `OrientModifier`
Add a new alternative `SimulationSpace::Local` for simulating particles
in the space defined by the `GlobalTransform` attached to the entity of
their `ParticleEffect`. This allows applying the effect's transform to
all particles after they are simulated. This is opposed to the default
global-space simulation offered by `SimulationSpace::Global` where
particles are "detached" from the effect when they spawn, and are
simulated in the global world space.

Add a new `OrientModifier` and its associated `OrientMode` enum,
allowing to orient the particles to face the camera and/or their
velocity, both when simulated in local and global space. This unified
modifier replaces both the `BillboardModifier` and the
`OrientAlongVelocityModifier`.

Bug: #180