Add `EffectShader` and use `naga_oil` to validate (#217)
Add a new internal `EffectShader` type containing all the configured shaders of
an effect, and an `EffectShaderSource` containing the generated source code for
an effect before it's specialized.

Add `naga_oil` as a dev-only dependency, to validate the generated code in
tests. This allows specializing the shader code (replace the `#if` _etc._)
without having to rely on the entire specialization pipeline feature of Bevy.