Customizable image sampling mapping
Add a new `ImageSamplingMapping` enum to customize the way a sample of
the image of a `ParticleTextureModifier` is mapped to and modulated with
the particle's base color.

The previous behavior was always using the Red channel of the texture as
an opacity mask, modulating the sample's Alpha component only. This
legacy behavior can be restored with the `ModulateOpacityFromR` value.

The new default is `Modulate`, which corresponds to a full
component-wise multiply of all components of the image sample by the
particle's base color.