Split runtime properties into `EffectProperties` (#256)
Split the dynamic runtime storage of property values away from
`CompiledParticleEffect`, which is an implementation detail, into a new
`EffectProperties` component. This component manages access to property
values by the user, as well as automated synchronization with the
underlying `EffectAsset` when the list of declared properties changes.

This change adds a new `EffectSystems::UpdatePropertiesFromAsset` set
during which the automated property update is performed. The set runs in
Bevy's own `UpdateAssets` schedule, before `Update`.

As a consequence, change detection on properties now works as expected,
and properties are only (re-)uploaded to GPU when the new
`EffectProperties` component changed, instead of every frame as this was
the case before this change. The component is also reflected, allowing
reflection-based access and mutations.