Add `OrientModifier::rotation`
Add an optional in-plane rotation to the `OrientModifier` to allow
rotating the particles within their oriented plane.

Add 4 new scalar float attributes `F32_0` to `F32_3` which have no
specific meaning, and can be used for anything. Use one in the
`billboard.rs` example to store a per-particle initial rotation, which
is later read back and used to set `OrientModifier::rotation` and give
some visual variations to the effect.

Change `Attribute::ALL` to private; replace with `Attribute::all()`.
This makes it a non-breaking change to add more attributes or reorder
them in the future.

Fixes #258