Handle attribute pointer in functions
Ensure that the generated code within a function, which generally
accesses the particle attribute struct via a pointer, is emitted with a
pointer indirection.

This is a partial fix, as `make_fn()` could be called with a non-pointer
attribute struct, which would break with the reverse error. However for
now it's only called in the update context with a pointer.

Fixes #275