Split render orientation and size
Split the calculation of the particle orientation (generally, from
`OrientModifier`) and its size (_e.g._ from `SizeOverLifetimeModifier`)
in such a way they become independent, and we can mix and match
orientations with screen space size or simulation space size.

Fix screen space size to avoid a perspective divide, which was
effectively breaking the depth-independent property of the size.

Fixes #269