Make `BinaryOperator::UniformRand` work with vectors.
The documentation claims that `UniformRand` works with vectors, but it
actually only works with scalars. This makes it hard to make particles
with random colors, especially with the lack of dynamic `vec4`
constructors.

This commit fixes the problem for float vectors only. It's still broken
with matrices and non-float scalars, but those are less important.