Implement particle groups, enabling simple particle trails.
This commit essentially allows multiple effects to share the same
underlying buffer. Such sub-effects are known as *particle groups*. An
effect may have up to 32 particle groups, each with its own capacity.
Every modifier specifies which particle groups it affects. Spawners
always spawn into particle group 0; this restriction was implemented for
simplicity and can be lifted later.

A new modifier has been added, `CloneModifier`. This allows particles to
be duplicated into a group on a fixed rate. By cloning particles into a
group, particle trails can be created.

It's intended that a node graph can be layered on top of particle groups
in the future. Particle groups are the most efficient way I can think of
to implement node graphs in any case, so this seems to me to be a step
toward a node-based editor.

Currently, init modifiers don't run on cloned particles. Their age is,
however, reset to 0 if present, as this is typically the most useful
behavior.

Internally, each particle group receives its own update and render
invocation. This reduces branching in both shaders, simplifies much of
the bookkeeping, and enables ribbon rendering for trails in the future
(needed for weapon effects, etc.)

The new `worms` example illustrates how trails can be produced with this
groups feature.