Add a global time scale for all effects (#267)
Add a new `Time<EffectSimulation>` clock which is used instead of the default
Bevy `Time` clock to tick effects. The clock "inherits" from the Bevy virtual
clock `Time<Virtual>`, such that the effective clock speed is the product of
both the Bevy virtual clock and this Hanabi effect simulation clock. As a
result, if any of those two clocks is paused, the effect simulation is paused.
This new clock allows manipulating the effect simulation time independently of
Bevy's own time if needed, while falling back by default on following Bevy's
clock.