Fix rendering of multiple effects
Fix the rendering of multiple effects, which was broken due to the use
of a `HashMap<>` keyed by the cached render pipeline ID, which is not
unique by definition (the cache is here to prevent duplicates, and reuse
a shader across effects and groups if the code is the same).

The core issue comes from the fact `EffectBatch` used to drive
rendering, mapping 1:1 with entities spawned in the render world to emit
draw calls. With the grouping of multiple draw calls per
`EffectBatches` (renamed from `EffectBatch`), each draw call of a group
inside an effect shares the same `EffectBatches` so has lost access to
its own unique group index.

To solve this, add a new `EffectDrawBatch` component, unique per draw
call, and containing the group index. For simplicity, this new component
references the existing `EffectBatches`, so we don't need to change
everything and we can share the data across groups. This adds one level
of indirection in the draw call, but hopefully ECS is fast enough for
the low number of calls we typically have. We can revisit later.