Move properties into `Module`
Move the declaration of properties from `EffectAsset` into `Module`.
Properties are only used in the context of expressions owned by a
`Module`, so it makes more sense for them to be owned by the `Module`
itself. This also forces them to be declared upfront with a type, which
enables adding type checking during expression building (not in this
change) instead of generating invalid shader code and only observing
issue when actually spawning an effect.