Add a modifier to make particles round. (#283)
This commit adds a `RoundModifier`, which allows particles to be made
into [squircle]s without having to use a texture. Squircles are like
rounded rectangles but are cheaper to evaluate. A configurable
`roundness` parameter allows the roundness to vary smoothly from a sharp
rectangle to a perfect ellipse.

Given x and y from (-1, 1), the equation of the shape of the particle is
|x|ⁿ + |y|ⁿ = 1, where n = 2 / `roundness`.

The `firework` and `portal` examples have been updated to use the
roundness modifier, as rounded particles generally look better than
square ones in these scenarios.

[squircle]: https://en.wikipedia.org/wiki/Squircle

---------

Co-authored-by: Jerome Humbert <djeedai@gmail.com>