Use strong handle in `CompiledParticleEffect`
This fixes a panic in the extract code, where the `EffectAsset` was
unloaded after the particle effect was compiled but before the
extraction finished. The idea of using a weak handle was to avoid
preventing the asset from unloading, but because
`CompiledParticleEffect` will anyway observe any change on the source
`ParticleEffect` and its handle, clearing that handle will make
`CompiledParticleEffect` also clear its own handle, even if strong,
which will result in the asset being "freed" from references held by
Hanabi.

Fixes #289