Cache bind group across frames (#307)
Increase the rendering performance by caching most bind groups across
frames, and only regenerate them when any of the underlying buffers was
re-allocated. Profiling with Tracy shows about 1ms saved on the `2d`
example modified as per #272 (10'000 entities, 100 particle effects).

- The visibility check of entities now uses a `FixedBitSet`. There's no
  notable performance difference in itself in the above test, but using
  a bitfield saves some memory. The actual saving comes from using a
  `Local` parameter, which caches the bitset across system runs and
  prevent allocation/free each frame.
- Move `EffectCache` as a standalone resource on the render world. This
  is mostly for convenience to prevent issues with the borrow checked,
  as the `EffectsMeta` contains too many things.
- In `EffectBuffer`, merge the init and update layouts and bind groups
  for the particle buffer and associated buffers. They were duplicated
  for no good reason, as the bound resources are the same. Also cache
  the bind groups inside the `EffectBuffer` itself instead of using a
  separate `HashMap`.
- Move the creation of bind groups into a new `prepare_bind_groups()`
  system running in the `EffectSystems::PrepareBindGroups` set, itself
  part of the `RenderSet::PrepareBindGroups`. This aligns Hanabi to the
  standard Bevy rendering architecture, ensuring that bind groups are
  created last after any buffer might have been re-allocated.

This change adds a new `trace` feature adding `info_span!()` trace
annotations when active. This allows profiling Hanabi, for example with
one of the various methods described in the Bevy documentation. To
profile with Tracy for example, run _e.g._:

```sh
cargo r --example 2d --no-default-features --release
--features="bevy/bevy_winit bevy/bevy_sprite 2d trace bevy/trace_tracy"
```

Note that this requires changing the default log level to Info,
otherwise the trace annotations are not emitted. This is the same for
Bevy and Wpgu.

Fixes #272