Fix effect properties ignored (#312)
Fix effect properties being ignored because there's no
`EffectProperties` component (and no code to auto-spawn it, despite the
documentation stating otherwise). Require the component instead to be
added always, and add it to `ParticleEffectBundle`. The alternative of
auto-spawning is rendered complex by the fact the decision requires
inspected the `EffectAsset`'s `Module` looking for properties, which
requires multiple levels of indirection, so would be rather expensive to
execute each frame on effects which doesn't require the component (so
keep not having it and require to be re-checked each frame).

Fixes #309