Handle padding for storage buffers
Handle the padding of the `GpuRenderGroupIndirect` structure so that it
complies with the GPU device requirements. This previously worked on
Windows because Vulkan or GPUs there require a 32-byte alignment which
was manually provided (hard-coded in Rust code). However this breaks on
Appled Mx GPUs with Metal, where the min alignment is 256 bytes.

Fixes #322