Fix shader binding bug (#356)
Fix a bug where the render shader attempts to bind the `Spawner` storage
buffer as read-only, but that struct is defined as containing an atomic
variable, which according to the WGSL specification requires write
access. Add a pipeline define `SPAWNER_READONLY` to redefine that struct
field as non-atomic, to allow read-only binding.