Add WebAssembly support
Add support for the WebAssembly target (`wasm`). This requires
disabling the `serde` feature, as the `typetag` used to serialize trait
objects is not available for that target.

Add a web page with all the examples, and move all the GIFs there.
Remove the list of examples from the README; instead point users to the
web demo at `examples/wasm/index.html`. This should be eventually
published live for even easier reference.

Temporarily copy `assets/` to `examples/wasm/` so that web examples can
directly access them when serving locally. This avoids complicated
multi-folder setups.

Fixes #41