Fixes for 1.79, add MSRV check in CI (#361)
* Revert "Fix new linter: remove `std::mem::` (#352)"

Rust 1.80 added `size_of()` and `align_of()` to the prelude. However
Bevy's MSRV (and therefore Hanabi's) is 1.79, so still require the
`std::mem::` qualification.

This reverts commit 0663370ccc37d831faeba1d3a049f9522df6957d.

* Remove `unused_qualifications` linter flag

This is necessary to build both with MSRV 1.79 (where `size_of()` needs
the `std::mem::` qualification) and 1.80 (where it was added to prelude
and doesn't need it). Otherwise there's always one version which fails
to build.

Fixes #360

* Add MSRV build to CI

* Run CI on hotfix/* branches

* Clippy fixes

* v0.12.2