Add support for texture sampling as expression
Add a new `TextureSampleExpr` to allow sampling a texture from an
expression. Make the `ParticleTextureModifier` use this. Change the
texture image to be hosted on a new `EffectMaterial` component, and add
`Module::add_texture()` to define a new texture slot (binding).

This is a primitive first version with several restrictions:
- Only color textures (those returning `vec4<f32>`), no depth.
- Only sampling via `textureSample()` in WGSL, no LOD or bias, no
  unsampled gather, _etc._
- No validation whatsoever about the coherence of the slots, the images
  on the material, and the expressions. If anything is inconsistent,
  this might produce runtime errors (fail to create shader) or simply
  render the wrong thing (no texture).

Fixes #355