Extract example app creation into utility function (#369)
Extract the creation of the Bevy app for all examples into a utility
function. The app creation is somewhat interesting to look at once, but
is otherwise very repetitive and clutters the examples with 20+ lines of
code, removing the focus on the actual specifics a particular example is
trying to demonstrate.

This change also ensures consistency across all examples, ensures
they're returning an error code on termination if any error occurs (for
CI), and make examples easier to maintain.