Optimize `BufferTable<T>` (#385)
Clean-up and optimize `BufferTable<T>`:
- Change row counts to `u32` type; this helps a bit readability.
- Fix doc link.
- Use `mapped_at_creation` to upload the newly added data when a new `Buffer`
  is created, instead of doing some buffer copies. This ensures all new rows
are copied in a single call, and makes better use of the WGPU staging buffer.

This change should not alter any behavior. It was originally a tentative
workaround for some bug with Metal debugging, but proved useless for that
purpose. However the minor optimization and code clean-up are nonetheless
relevant.