Fix panic in clone init pass
Fix a panic in the clone init pass due to some Vulkan validation error
when using ribbons and trails. The indirect dispatch of the clone init
pass used the wrong buffer offset, which in some cases was leading to an
out of bound access, caught by the validation layer.

Fixes #399