Start transitioning to retained render world
Start to transition the internal rendering implementation to leverage
the retained render world introduced in Bevy 0.15.

This change moves the `CachedEffect` and `DispatchBufferIndices` out of
the `EffectCache`, and convert them to components added to the render
world entity spawned and synchronized via `SyncToRenderWorld` with the
main world entity owning the effect instance.

As part of this change, split the addition and removal of new effect
instances into a separate `add_remove_effects()` system which runs
before the `prepare_effects()` system.

Also fix a small bug with opaque particles not rendered in some cases;
this is hypothetical, but `EffectShaderSource::generate()` was not
setting the `LayoutFlags::OPAQUE`.