Split properties into new `PropertyCache`
Split the properties from the `EffectCache` and move their resource
management into a new `PropertyCache`, which allocates GPU buffers to
store them. This clarifies ownership, and allows better batching (in
theory) as not all effects use properties.

In this change, unfortunately for simplicity we directly store the
offset into the property buffer for each instance into the effect batch,
meaning we again block batching multiple effects. This is hopefully
temporary (and not the only blocker, so doesn't introduce any
regression), but makes the change incremental and easier to handle for
now, while other clean-ups are still pending.