Invalidate more bind group on changes
Invalidate the bind groups for the simulation, containing the particle
group buffer, when that buffer is reallocated or when the group changes
position in the buffer. This was causing old buffers to be referenced
when the group(s) changed, creating artifacts in ribbon rendering in
particular.

Similarly, invalidate the bind groups referencing a property buffer when
that buffer is reallocated, for the same reasons. This fixes a crash
previously introduced by the new property buffer change.