Destroy buffers after use, fix bug in `BufferTable`
Always call `Buffer::destroy()` on the old buffer resource when a GPU
buffer is re-allocated. This ensures we get a runtime assertion if a
bind group is not recreated and attempt to use a stale buffer.

Fix a bug in `BufferTable` where the free list was not updated
correctly, which could in theory lead to issues once re-allocating new
elements.