Add spawner cycle count (#434)
Add `Spawner::cycle_count` to limit the number of times the spawning pattern repeats.

Rename `Spawner` to `SpawnerSettings` for clarity. The actual component triggering the spawning is `EffectSpawner`.

Rename `starts_imediately`, which didn't work, into `emit_on_start`. Clarify docs about the fact it's related to the component spawning in the ECS world, and it has no effect on spawners repeating infinitely.

Fixes #273