Fix bug in indexing the spawner params (#441)
The effect metadata was maintaining a stale index to the spawner, which
was incorrectly used in the indirect pass instead of using the same
effect index as when read a few lines above. The effect metadata can't
possibly maintain an index to the spawner because the former is lazily
uploaded (and in fact this resets the effect, so should be almost never)
while the latter is re-uploaded each frame and is always tightly packed
(we reallocated the spawners each frame since anyway they change each
frame).

Fixes #412