Add an `OnSpawn` condition for particle emission.
[Unity has this.]

An example use case is spawning a bunch of particles as clumps around a
circle. You make a hidden particle system A that uses the
`SetPositionCircleModifier` to spawn invisible particles around a
circle, then add an `EmitSpawnEventModifier` with an `OnSpawn` condition
that triggers another particle system B to spawn, say, 5 particles for
each particle in system A. Particle system B uses a
`SetPositionSphereModifier` and inherits the position from particle
system A. That way you get some nice clumps of particles all spaced
randomly around a circle.

[Unity has this.]: https://docs.unity3d.com/6000.0/Documentation/ScriptReference/ParticleSystemSubEmitterType.Birth.html