Make the number of particles to emit in a GPU event an expression instead of a constant. (#444)
This makes Hanabi match Unity, which supports this feature [as explained
here]. It allows for a variety of effects. For example, one could emit a
random number of particles on every frame instead of a fixed number. Or
one could emit a "burst" of particles after traveling some distance, by
conditionally setting the number of particles to spawn to 0.

[as explained here]: https://discussions.unity.com/t/gpuevent-trigger-once-at-birth/779939/2