Recreate the init fill dispatch bind group when the arguments buffer resizes.
This is currently not done, resulting in a crash when the arguments
buffer resizes, which typically happens when multiple particle effects
are spawned simultaneously.

There's a `todo!` that suggests that the original plan was to eagerly
invalidate the init fill dispatch buffers when the arguments buffer
changes size. I replaced this with usage of `CachedBindGroup` keyed off
the ID of the arguments buffer, which seems to be the pattern used
elsewhere in the code.

Closes #446.